fix(schemas, repository, service): Change fare data type to integer
Changed the data type of the fare field across various schemas, repositories, and services from float64 to int64.  This change is necessary to accurately represent the fare in Vietnamese Dong (VND), which doesn't use decimal values, and to ensure consistency with the database model.  The fare is now rounded to the nearest 1000 VND.  This also improves the precision of fare calculations in the map_repo.go file.

Authored-by: Lê Minh Trí <leminhtri11a1@gmail.com>

+++ infra/db/migration/migration_model.go
@@ -113,7 +113,7 @@ type Transaction struct {
 	Payer         User      `gorm:"foreignKey:PayerID"`
 	ReceiverID    uuid.UUID `gorm:"type:uuid"`
 	Receiver      User      `gorm:"foreignKey:ReceiverID"`
-	Amount        float64
+	Amount        int64     // in vnđ so cannot have decimal and 1000 is the smallest currency unit
 	PaymentMethod string    `gorm:"default:'cash'"`    // cash, momo
 	Status        string    `gorm:"default:'pending'"` // pending, completed, failed, refunded
 	RideID        uuid.UUID `gorm:"type:uuid"`
@@ -159,7 +159,7 @@ type RideOffer struct {
 	Rides                  []Ride  `gorm:"foreignKey:RideOfferID"`
 	StartTime              time.Time
 	EndTime                time.Time  // Time to end the ride (end time = start time + duration)
-	Fare                   float64    // Total price of the ride offer (to show to the hitchhiker)
+	Fare                   int64      // Total price of the ride offer (to show to the hitchhiker)
 	Waypoints              []Waypoint `gorm:"foreignKey:RideOfferID"`
 }
 
@@ -213,7 +213,7 @@ type Ride struct {
 	Status          string      `gorm:"default:'scheduled'"` // scheduled, ongoing, completed, cancelled
 	StartTime       time.Time
 	EndTime         time.Time
-	Fare            float64
+	Fare            int64             // Total price of the ride (to show to the hitchhiker, vnđ so cannot use float)
 	StartAddress    string            `gorm:"type:text"`
 	EndAddress      string            `gorm:"type:text"`
 	EncodedPolyline polyline.Polyline `gorm:"type:text"`
